+++ src/game/BattleGround/BattleGroundAV.h
@@ -670,7 +670,7 @@ class BattleGroundAV : public BattleGround
         void ChangeMineOwner(AVMineIds mineId, PvpTeamIndex newOwnerTeamIdx);
 
         // World state helpers
-        uint8 GetWorldStateType(uint8 state, PvpTeamIndex teamIdx) const { return teamIdx * BG_AV_MAX_STATES + state; }
+        uint8 GetWorldStateType(uint8 state, PvpTeamIndex teamIdx) const { return static_cast<uint16>(teamIdx) * static_cast<uint16>(BG_AV_MAX_STATES) + state; }
         void UpdateNodeWorldState(AVNodeIds node, uint32 newState);
 
         // Herald announcements
